+++ test/fixtures/codes/less.js
@@ -1,0 +1,9 @@
+module.exports = `@width: 10px;
+@height: @width + 10px;
+
+body {
+  background: red;
+  width: @width;
+  height: @height;
+}
+`