+++ test/fixtures/codes/ts.js
@@ -1,0 +1,1 @@
+module.exports = 'const x: number = 0; console.log(x)'