+++ test/fixtures/codes/umd.js
@@ -1,0 +1,6 @@
+module.exports = `
+import React from 'react'
+export default function () {
+  return (<div></div>)
+}
+`