+++ test/fixtures/react/components/index.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import 'reset-css/less/reset.less'
+
+export default function () {
+  return (
+    <div>component</div>
+  )
+}