+++ test/fixtures/ts/index.tsx
@@ -1,9 +1,6 @@
 import React from 'react'
 import { render } from 'react-dom'
-
 import { Hello } from './hello'
-
-console.log('?')
 
 render(
   <Hello compiler="TypeScript" framework="React" />,